ci: fix dry-run deploy failure in dependabot PRs
Dependabot PRs are treated as "fork PRs" for security reasons, meaning
they don't have access to the repo secrets. This caused all Dependabot
PRs to fail CI because AWS keys were blank for the dry-run deploy.

We can fix this by effectively making the AWS keys optional when the
dry-run option is used, and if they aren't present, not attempting the
`aws` command at all (but still going through the packaging process).