feat(packager): allow overwriting output file
This lets us more easily regenerate output files when bugs are fixed in
the packager.

Technically we don't have to adhere to the `ScheduledEvent` type, since
manually invoking from the CLI is *not* a scheduled event and there are
no restrictions on the format of data sent this way. Stashing the option
in `detail` is just a convenient way to keep the existing types as-is.